+++ lib/ood_core/job/task_status.rb
@@ -1,0 +1,24 @@
+module OodCore
+  module Job
+    class TaskStatus
+      attr_reader :id
+      attr_reader :status
+
+      def initialize(id:, status:, **_)
+        @task_id = id
+        @status = OodCore::Job::Status.new(state: status)
+      end
+
+      def to_h
+        {
+          :id => @id,
+          :status => @status
+        }
+      end
+
+      def ==(other)
+        self.to_h == other.to_h
+      end
+    end
+  end
+end